Move nrf 15.4 radio channel enum to HIL
- Using this enum insures only valid 15.4 radio channels are passed
