+++ kernel/src/hil/radio.rs
@@ -104,7 +104,7 @@ pub trait RadioConfig<'a> {
     fn set_address_long(&self, addr: [u8; 8]);
     fn set_pan(&self, id: u16);
     fn set_tx_power(&self, power: i8) -> Result<(), ErrorCode>;
-    fn set_channel(&self, chan: u8) -> Result<(), ErrorCode>;
+    fn set_channel(&self, chan: RadioChannel);
 }
 
 pub trait RadioData<'a> {
@@ -117,4 +117,74 @@ pub trait RadioData<'a> {
         spi_buf: &'static mut [u8],
         frame_len: usize,
     ) -> Result<(), (ErrorCode, &'static mut [u8])>;
+}
+
+#[derive(PartialEq, Debug, Copy, Clone)]
+pub enum RadioChannel {
+    DataChannel11 = 5,
+    DataChannel12 = 10,
+    DataChannel13 = 15,
+    DataChannel14 = 20,
+    DataChannel15 = 25,
+    DataChannel16 = 30,
+    DataChannel17 = 35,
+    DataChannel18 = 40,
+    DataChannel19 = 45,
+    DataChannel20 = 50,
+    DataChannel21 = 55,
+    DataChannel22 = 60,
+    DataChannel23 = 65,
+    DataChannel24 = 70,
+    DataChannel25 = 75,
+    DataChannel26 = 80,
+}
+
+impl RadioChannel {
+    /// Returns the u8 value of the channel, which is the IEEE 802.15.4 channel
+    pub fn get_channel_index(&self) -> u8 {
+        match *self {
+            RadioChannel::DataChannel11 => 11,
+            RadioChannel::DataChannel12 => 12,
+            RadioChannel::DataChannel13 => 13,
+            RadioChannel::DataChannel14 => 14,
+            RadioChannel::DataChannel15 => 15,
+            RadioChannel::DataChannel16 => 16,
+            RadioChannel::DataChannel17 => 17,
+            RadioChannel::DataChannel18 => 18,
+            RadioChannel::DataChannel19 => 19,
+            RadioChannel::DataChannel20 => 20,
+            RadioChannel::DataChannel21 => 21,
+            RadioChannel::DataChannel22 => 22,
+            RadioChannel::DataChannel23 => 23,
+            RadioChannel::DataChannel24 => 24,
+            RadioChannel::DataChannel25 => 25,
+            RadioChannel::DataChannel26 => 26,
+        }
+    }
+}
+
+impl TryFrom<u8> for RadioChannel {
+    type Error = ();
+    /// Returns the RadioChannel for the given u8 value, which is the IEEE 802.15.4 channel
+    fn try_from(val: u8) -> Result<RadioChannel, ()> {
+        match val {
+            11 => Ok(RadioChannel::DataChannel11),
+            12 => Ok(RadioChannel::DataChannel12),
+            13 => Ok(RadioChannel::DataChannel13),
+            14 => Ok(RadioChannel::DataChannel14),
+            15 => Ok(RadioChannel::DataChannel15),
+            16 => Ok(RadioChannel::DataChannel16),
+            17 => Ok(RadioChannel::DataChannel17),
+            18 => Ok(RadioChannel::DataChannel18),
+            19 => Ok(RadioChannel::DataChannel19),
+            20 => Ok(RadioChannel::DataChannel20),
+            21 => Ok(RadioChannel::DataChannel21),
+            22 => Ok(RadioChannel::DataChannel22),
+            23 => Ok(RadioChannel::DataChannel23),
+            24 => Ok(RadioChannel::DataChannel24),
+            25 => Ok(RadioChannel::DataChannel25),
+            26 => Ok(RadioChannel::DataChannel26),
+            _ => Err(()),
+        }
+    }
 }
